chore: update Tabs internal structure to allow type generation
